Add files via upload
This repository contains the implementation of the classic Simon game. The game is a simple memory game where the player must remember and repeat a sequence of colors.

The game starts when the user presses a key. The game will then show the first color in the sequence and the user must click the correct color. If the user's answer is correct, the game will show the next color in the sequence. The user must then click the colors in the correct order. The sequence gets longer and faster with each level, challenging the user's memory and reaction speed.

The game is implemented using HTML, CSS, and JavaScript. The HTML file sets up the game's interface, including the start screen and the colored buttons. The CSS file styles the game's interface, and the JavaScript file implements the game's logic.

The game also includes audio feedback. Each color has a corresponding sound that plays when the color is shown in the sequence or clicked by the user. There is also a sound that plays when the user's answer is wrong.

If the user's answer is wrong, the game is over and the user can start over by pressing any key. The game's state is reset and the user can try to beat their high score.